Bump retrofit and remove retrofitCoroutinesAdapter dependency (#1001)
* Bump kotlin and coroutine libraries

* Bump retrofit version
Remove retrofit-coroutine adapter library (deprecated: https://github.com/JakeWharton/retrofit2-kotlin-coroutines-adapter#kotlin-coroutine-adapter-deprecated)
Adapt code to introduce suspend operator

* bump coroutines version!

* remove unnecessary annotations

* Fixes compilation error due to DuplicateRelativeFileException

* fix failing test related to withTimeoutOrNull (Job has not completed yet)
