+++ app/build.gradle
@@ -110,14 +110,13 @@ ext {
     legacySupport = "1.0.0"
     coreTesting = "2.1.0"
     testRunner = "1.3.0"
-    coroutines = "1.3.5"
-    retrofitCoroutinesAdapter = "0.9.2"
+    coroutines = "1.4.2"
     webkit = "1.3.0"
     referrerLibrary = "1.1.2"
 
     junit = "4.12"
     dagger = "2.27"
-    retrofit = "2.8.1"
+    retrofit = "2.9.0"
     ankoVersion = "0.10.4"
     glide = "4.11.0"
     lottieVersion = "3.4.0"
@@ -141,7 +140,12 @@ dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines"
-    androidTestImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$coroutines"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines"
+    androidTestImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:$coroutines") {
+        // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
+        // conflicts with mockito due to direct inclusion of byte buddy
+        exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
+    }
 
     implementation "androidx.appcompat:appcompat:$androidX"
     implementation "com.google.android.material:material:$materialDesign"
@@ -154,7 +158,6 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-moshi:$retrofit"
     implementation "com.squareup.retrofit2:adapter-rxjava2:$retrofit"
     implementation "com.squareup.retrofit2:converter-scalars:$retrofit"
-    implementation "com.jakewharton.retrofit:retrofit2-kotlin-coroutines-adapter:$retrofitCoroutinesAdapter"
     implementation "io.reactivex.rxjava2:rxjava:$rxJava"
     implementation "io.reactivex.rxjava2:rxandroid:$rxAndroid"
     implementation "com.jakewharton.timber:timber:$timber"
