+++ app/src/main/java/com/duckduckgo/app/feedback/api/FeedbackService.kt
@@ -16,7 +16,6 @@
 
 package com.duckduckgo.app.feedback.api
 
-import retrofit2.Call
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.POST
@@ -25,7 +24,7 @@ interface FeedbackService {
 
     @FormUrlEncoded
     @POST("/feedback.js?type=app-feedback")
-    fun submitFeedback(
+    suspend fun submitFeedback(
         @Field("reason") reason: String = REASON_GENERAL,
         @Field("rating") rating: String,
         @Field("category") category: String?,
@@ -38,7 +37,7 @@ interface FeedbackService {
         @Field("manufacturer") manufacturer: String,
         @Field("model") model: String,
         @Field("atb") atb: String
-    ): Call<Void>
+    )
 
     companion object {
         const val REASON_GENERAL = "general"
